# Conflicts:
#	android/src/main/java/app/saboten/androidApp/ui/list/PostSelectItem.kt
#	android/src/main/java/app/saboten/androidApp/ui/screens/post/PostDetailScreen.kt

+++ server/src/main/java/backend/controller/CommentController.java
@@ -24,7 +24,7 @@ import org.springframework.web.bind.annotation.*;
 
 @Version1RestController
 @RequiredArgsConstructor
-public class CommentController {
+public class  CommentController {
 
     private final UserService userService;
     private final PostService postService;
@@ -47,7 +47,7 @@ public class CommentController {
         UserEntity userEntity = getUser();
         PostEntity postEntity = postService.findPost(postId);
         UserResponse userResponse = new UserResponse(userEntity.getUserId(), userEntity.getNickname(), userEntity.getUserImage());
-        Integer voteSelectEntity = voteSelectService.findVoteSelectResult(userEntity,postEntity);
+        Long voteSelectEntity = voteSelectService.findVoteSelectResult(userEntity,postEntity);
         String text = commentCreateRequest.getText();
         CommentEntity comment = commentService.create(userEntity, postEntity,text);
         CommentResponse commentResponse = new CommentResponse(comment.getCommentId(), comment.getCommentText(),
